requirements.txt
streamlit
pytest
pypdf
python-docx
spacy
rapidfuzz
pandas
